fix(pr-323): reduce complexity and broaden DeepSource exclusions
Split monitor page hooks and WMS playback helpers for CodeScene, and
exclude monitor-only paths from DeepSource hygiene/doc noise.
